+++ packages/react/src/krillswitch.tsx
@@ -15,7 +15,7 @@ import {
   mergeIntoManifest,
   requestEvaluation,
   type WidenManifest,
-} from "./evaluation";
+} from "./evaluation.js";
 
 export interface FeatureFlagProviderProps<
   M extends FlagManifest = FlagManifest,